clase2/requirements.txt
flask
python-dotenv
flask-restful